pyproject.toml
[tool.poetry]
name = "getticket"
version = "0.1.0"
description = ""
authors = ["Keunreol Park <ritter4u@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
